# docs/openapi-week3.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenGuard Kids API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/device/event-state/ack": {
      "post": {
        "summary": "Acknowledge Erasure",
        "operationId": "acknowledge_erasure_api_device_event_state_ack_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerationAck"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/device/activity-events": {
      "post": {
        "summary": "Ingest",
        "operationId": "ingest_api_device_activity_events_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventBatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/children/{child_id}/activity-events": {
      "get": {
        "summary": "History",
        "operationId": "history_api_children__child_id__activity_events_get",
        "parameters": [
          {
            "name": "child_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Child Id"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/children/{child_id}/activity-data/deletion": {
      "get": {
        "summary": "Erasure Status",
        "operationId": "erasure_status_api_children__child_id__activity_data_deletion_get",
        "parameters": [
          {
            "name": "child_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Child Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/children/{child_id}/activity-data": {
      "delete": {
        "summary": "Erase",
        "operationId": "erase_api_children__child_id__activity_data_delete",
        "parameters": [
          {
            "name": "child_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Child Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/children/{child_id}/filtering": {
      "get": {
        "summary": "Parent Filtering",
        "operationId": "parent_filtering_api_children__child_id__filtering_get",
        "parameters": [
          {
            "name": "child_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Child Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Set Filtering",
        "operationId": "set_filtering_api_children__child_id__filtering_put",
        "parameters": [
          {
            "name": "child_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Child Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FilterInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ActivityPayload": {
        "properties": {
          "ts": {
            "type": "number",
            "minimum": 0.0,
            "title": "Ts"
          },
          "device_id": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Device Id"
          },
          "child_id": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Child Id"
          },
          "type": {
            "type": "string",
            "enum": [
              "app_start",
              "app_stop",
              "domain_query",
              "blocked_app",
              "blocked_domain",
              "quota_warning",
              "locked",
              "unlock_request",
              "override_granted"
            ],
            "title": "Type"
          },
          "subject": {
            "type": "string",
            "maxLength": 253,
            "minLength": 1,
            "title": "Subject"
          },
          "duration_sec": {
            "type": "integer",
            "maximum": 7776000.0,
            "minimum": 0.0,
            "title": "Duration Sec"
          },
          "policy_id": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Policy Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "ts",
          "device_id",
          "child_id",
          "type",
          "subject",
          "duration_sec",
          "policy_id"
        ],
        "title": "ActivityPayload"
      },
      "AppRule": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "pattern": "^[^/\\\\\\r\\n]+$",
            "title": "Name"
          },
          "sha256": {
            "type": "string",
            "pattern": "^[a-fA-F0-9]{64}$",
            "title": "Sha256"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "sha256"
        ],
        "title": "AppRule"
      },
...
      "EventBatch": {
        "properties": {
          "events": {
            "items": {
              "$ref": "#/components/schemas/EventEnvelope"
            },
            "type": "array",
            "maxItems": 100,
            "minItems": 1,
            "title": "Events"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "events"
        ],
        "title": "EventBatch"
      },
      "EventEnvelope": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "generation": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Generation"
          },
          "event": {
            "$ref": "#/components/schemas/ActivityPayload"
          },
          "explanation": {
            "$ref": "#/components/schemas/Explanation"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "generation",
          "event"
        ],
        "title": "EventEnvelope"
      },
      "Explanation": {
        "properties": {
          "reason": {
            "type": "string",
            "maxLength": 300,
            "title": "Reason",
            "default": ""
          },
          "rule_author": {
            "type": "string",
            "maxLength": 80,
            "title": "Rule Author",
            "default": "Phụ huynh"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "Explanation"
      },
      "FilterInput": {
        "properties": {
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": false
          },
          "app_mode": {
            "type": "string",
            "enum": [
              "blocklist",
              "allowlist"
            ],
            "title": "App Mode",
            "default": "blocklist"
          },
          "app_allow": {
            "items": {
              "$ref": "#/components/schemas/AppRule"
            },
            "type": "array",
            "maxItems": 200,
            "title": "App Allow"
          },
          "app_block": {
            "items": {
              "$ref": "#/components/schemas/AppRule"
            },
            "type": "array",
            "maxItems": 200,
            "title": "App Block"
          },
          "domain_mode": {
            "type": "string",
            "enum": [
              "blocklist",
              "allowlist"
            ],
            "title": "Domain Mode",
            "default": "blocklist"
          },
          "domain_allow": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 500,
            "title": "Domain Allow"
          },
          "domain_block": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 500,
            "title": "Domain Block"
          },
          "safety_domains": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Safety Domains"
          },
          "expected_version": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Expected Version"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "expected_version"
        ],
        "title": "FilterInput"
      },
      "GenerationAck": {
        "properties": {
          "generations": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "maxProperties": 100,
            "title": "Generations"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "generations"
        ],
        "title": "GenerationAck"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}